requirements.txt
numpy
pandas
opencv-python
pillow
easyocr
torch
torchvision
transformers
scikit-learn